Remove conversion warning from src/index directory 	modified:   src/index/chain/MonotoneChain.cpp 	modified:   src/index/quadtree/DoubleBits.cpp 	modified:   src/index/quadtree/NodeBase.cpp

+++ src/index/quadtree/DoubleBits.cpp
@@ -118,7 +118,7 @@ int
 DoubleBits::getExponent() const
 {
 #if ASSUME_IEEE_DOUBLE
-	return biasedExponent()-EXPONENT_BIAS;
+	return static_cast<int>(biasedExponent() - EXPONENT_BIAS);
 #else
 	if ( x <= 0 ) return 0; // EDOM || ERANGE
 	return (int)((log(x)/log(2.0))+(x<1?-0.9:0.00000000001));
